Change the bool value returned by scheduleCall in Schedule to bytes
Value returned by the scheduleCall() function in Schedule used to be
bool, signifying the successful execution of the function. This commit
changes it to the bytes value that is equal to task_id of the task
created by calling the function. This way the smart contracts calling
the Schedule smart contract are able to provide even more automation in
regards to interacting with the Schedule smart contract.